pyproject.toml
[tool.poetry]
name = "django-faucet"
version = "0.1.0"
description = ""
authors = ["Alex Kvasha <alex.kvasha.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
djangorestframework = "^3.15.2"
gunicorn = "^23.0.0"
web3 = "^7.8.0"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
